ignore linter

+++ pkg/fftypes/uuid.go
@@ -96,9 +96,11 @@ func (u *UUID) HashBucket(buckets int) int {
 	if u == nil || buckets <= 0 {
 		return 0
 	}
-	// Take the last random 4 bytes and mod it against the bucket count to generate
-	// a deterministic hash bucket allocation for the UUID V4
-	return int(binary.BigEndian.Uint16((*u)[8:])) % buckets
+	// Explicitly cast `buckets` to uint64 to match the type used in the modulo operation
+	// and ensure the result is safely converted back to int.
+
+	// #nosec G115 - Safe because `buckets` is calculated using modulo operation
+	return int(binary.BigEndian.Uint64((*u)[8:]) % uint64(buckets))
 
 }
